Another small refactor, move the main handler in __init__.py
