+++ scripts/idl/zapxml/handlers/handlers.py
@@ -579,15 +579,3 @@ class ConfigurationHandler(BaseHandler):
             return GlobalHandler(self.context)
         else:
             return BaseHandler(self.context)
-
-
-class ZapXmlHandler(BaseHandler):
-    def __init__(self, context: Context, idl: Idl):
-        super().__init__(context)
-        self._idl = idl
-
-    def GetNextProcessor(self, name, attrs):
-        if name.lower() == 'configurator':
-            return ConfigurationHandler(self.context, self._idl)
-        else:
-            return BaseHandler(self.context)
